add: descrição para algoritmos
